requirements.txt
requests~=2.31.0
Js2Py~=0.74
beautifulsoup4~=4.12.2
cloudscraper~=1.2.71
websocket-client~=1.6.4
appdirs~=1.4.4
pydantic~=2.5.2